requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
PyYAML==6.0
tk==0.1.0
toml==0.10.2
